+++ web/Controllers/EditorsController.cs
@@ -117,6 +117,7 @@ namespace ASC.Api.Web.Help.Controllers
                 "OwnCloud",
                 "Plone",
                 "Plugins",
+                "ProtectingRanges",
                 "React",
                 "Redmine",
                 "Rename",
@@ -428,6 +429,11 @@ namespace ASC.Api.Web.Help.Controllers
         }
 
         public ActionResult Plugins()
+        {
+            return View();
+        }
+
+        public ActionResult ProtectingRanges()
         {
             return View();
         }